+++ public/ide/modules/ide/js/IDEController.js
@@ -34,13 +34,13 @@ console.log = (function (console) {
 
 var logs=[];
 ide.controller('IDECtrl', [ '$scope','$timeout','$http','$interval', 'ProjectService', 'GlobalService', 'Preference', 'ResourceService', 'TagService', 'TemplateProvider','TimerService','UserTypeService','WidgetService','NavModalCANConfigService',
-    function ($scope,$timeout,$http,$interval,
+    'socketIOService',function ($scope,$timeout,$http,$interval,
                                     ProjectService,
                                     GlobalService,
                                     Preference,
                                     ResourceService,
                                     TagService,
-                                    TemplateProvider,TimerService,UserTypeService,WidgetService,NavModalCANConfigService) {
+                                    TemplateProvider,TimerService,UserTypeService,WidgetService,NavModalCANConfigService,socketIOService) {
 
     ideScope=$scope;
     $scope.ide={
@@ -278,6 +278,13 @@ ide.controller('IDECtrl', [ '$scope','$timeout','$http','$interval', 'ProjectSer
     }
 
     function loadFromContent(data,id) {
+        //若是分享的工程，则需要开启socket
+        console.log('data.shared',data.shared);
+        if(!!data.shared){
+            if(!socketIOService.getSocket()){
+                initSocketIO(data.userId);
+            }
+        }
         //change html title to name
         var name = data&&data.name||''
         document.title = '工程编辑-'+name
@@ -453,7 +460,6 @@ ide.controller('IDECtrl', [ '$scope','$timeout','$http','$interval', 'ProjectSer
             $scope.ide.loaded=true;
             window.spinner && window.spinner.hide();
             // intervalSave();
-
         },200)
     }
 
@@ -1156,4 +1162,141 @@ ide.controller('IDECtrl', [ '$scope','$timeout','$http','$interval', 'ProjectSer
         }
 
     };
+
+    /**
+     * 初始化socket
+     */
+    $scope.wrapperForCoop = false;
+    $scope.projectOwner = false;
+    var inCharge = false;
+    function initSocketIO(ownerId){
+        socketIOService.createSocket('',function(data){
+
+            console.log('you have connect');
+
+            //capture current users in room
+            socketIOService.on('connect:success',function(allUsers,currentUser){
+                // console.log('connect:success',allUsers);
+                socketIOService.setRoomUsers(allUsers);
+                $scope.currentUser = currentUser;
+                inCharge = allUsers[0]&&(allUsers[0].id===$scope.currentUser.id);
+                if(!inCharge){
+                    console.log('wenerId',ownerId,'currId',currentUser.id);
+                    $scope.projectOwner = (ownerId===currentUser.id);
+                    $scope.wrapperForCoop = true;
+                    $scope.currentUsers = socketIOService.getRoomUsers();
+                }
+            });
+
+
+            //capture new user enter this room
+            socketIOService.on("user:enter",function(data){
+                toastr.info('用户 '+data.username +'加入');
+                $scope.$apply(function(){
+                    $scope.currentUsers = socketIOService.addUserInRoom(data);
+                    // console.log('add user',$scope.currentUsers);
+                })
+            });
+
+            //capture user leave this room
+            socketIOService.on("user:leave",function (data) {
+                toastr.info('用户 '+data.username+' 已离开');
+                $scope.$apply(function(){
+                    $scope.currentUsers = socketIOService.deleteUserInRoom(data);
+                });
+
+                //check currentUser have right to edit project
+                // console.log('inCharge',inCharge);
+                if(!inCharge){
+                    inCharge = $scope.currentUsers[0]&&($scope.currentUsers[0].id===$scope.currentUser.id);
+                    $timeout(function(){
+                        if(inCharge){
+                            alert('您已获得编辑工程的权限，即将重新加载工程');
+                            window.spinner&&window.spinner.show();
+                            $scope.wrapperForCoop = false;
+                            loadStep = 0;
+                            readProjectData();
+                        }
+                    },1500)
+                }
+            });
+
+
+            //capture room close
+            socketIOService.on('room:close',function(){
+                toastr.warning('管理员已经关闭共享，页面即将关闭');
+                socketIOService.closeSocket(function(){
+                    setTimeout(function(){
+                        closeWebPage();
+                    },1000)
+                })
+            })
+        });
+
+    }
+
+    $scope.openSimulator = function(){
+        $scope.$broadcast('OpenSimulator')
+    };
+
+    $scope.cancelShare = function(){
+        if(confirm('强制取消将对正在编辑的工程产生影响，确定取消分享？')){
+            var arr = window.location.href.split('/');
+            var id = arr[arr.length-2];
+            console.log('id',id);
+            $http({
+                method:'POST',
+                url:'/project/'+id+'/share',
+                data:{
+                    share:false
+                }
+            })
+            .success(function(data,status,xhr){
+                socketIOService.emit('room:close');
+                socketIOService.closeSocket();
+                toastr.info('取消成功,工程即将重新加载!');
+                setTimeout(function(){
+                    window.spinner&&window.spinner.show();
+                    $scope.wrapperForCoop = false;
+                    loadStep = 0;
+                    readProjectData();
+                },1500);
+
+            })
+            .error(function(err){
+                console.log(err)
+            });
+        }
+    };
+
+    $scope.$on('createSocketIO',function () {
+        console.log('open share then create socketIO');
+        initSocketIO();
+    });
+
+    $scope.$on('closeSocketIO',function(){
+        console.log('close share then close socketIO');
+        socketIOService.emit('room:close');
+        socketIOService.closeSocket();
+    });
+
+
+    function closeWebPage(){
+        if (navigator.userAgent.indexOf("MSIE") > 0) {
+            if (navigator.userAgent.indexOf("MSIE 6.0") > 0) {
+                window.opener = null;
+                window.close();
+            } else {
+                window.open('', '_top');
+                window.top.close();
+            }
+        }
+        else if (navigator.userAgent.indexOf("Firefox") > 0) {
+            window.location.href = 'about:blank ';
+        } else {
+            window.opener = null;
+            window.open('', '_self', '');
+            window.close();
+        }
+    }
 }]);