+++ public/ide/modules/ide/js/parts/nav/navController.js
@@ -109,6 +109,7 @@ ide.controller('NavCtrl', ['$scope', '$timeout',
             ProjectService.getProjectTo($scope);
             $scope.$on('NavStatusChanged', onNavStatusChanged);
 
+            $scope.$on('OpenSimulator',$scope.component.tool.play)
             // setInterval(function () {
             //     saveProject();
             // }.bind(this),5*60*1000)
@@ -1130,6 +1131,7 @@ ide.controller('NavCtrl', ['$scope', '$timeout',
                 animation: $scope.animationsEnabled,
                 templateUrl: 'shareModal.html',
                 controller: 'shareModalCtl',
+                scope:$scope,
                 size: 'md',
                 resolve: {
                     id:function () {
@@ -1144,7 +1146,7 @@ ide.controller('NavCtrl', ['$scope', '$timeout',
                 //process save
                 generateDataFile(result.format);
             }, function () {
-                console.log('Modal dismissed at: ' + new Date());
+
             });
         }
 
@@ -1454,10 +1456,8 @@ ide.controller('shareModalCtl',['$rootScope','$scope','$uibModalInstance','$http
     }
 
 
-
-
     $scope.toggleShare = function () {
-        $scope.processing = true
+        $scope.processing = true;
         $http({
             method:'POST',
             url:'/project/'+id+'/share',
@@ -1470,6 +1470,12 @@ ide.controller('shareModalCtl',['$rootScope','$scope','$uibModalInstance','$http
             $scope.shareInfo.sharedKey = data.sharedKey
             $scope.processing = false
             $scope.message = ''
+
+            if(data.shared){
+                $scope.$emit('createSocketIO');
+            }else{
+                $scope.$emit('closeSocketIO');
+            }
         })
         .error(function(err){
             console.log(err)
@@ -1479,7 +1485,7 @@ ide.controller('shareModalCtl',['$rootScope','$scope','$uibModalInstance','$http
     }
 
     $scope.cancel = function () {
-        $uibModalInstance.dismiss('cancel');
+        $uibModalInstance.dismiss($scope.shareInfo.shared);
     };
 }]);