+++ routes/index.js
@@ -112,6 +112,9 @@ router.route('/user/login')
 router.route('/user/loginAPI')
 .post(loginAPI.post);
 
+router.route('/user/logininfo')
+    .get(loginAPI.getLoginInfo)
+
 //logout
 router.route('/user/logout')
 .get(route_login.logout);
@@ -168,10 +171,13 @@ router.route('/project/:id/basicinfo')
     .post(projectInfo.updateProject);
 
 router.route('/project/:id/backuplist')
-    .get(projectInfo.getBackupList)
+    .get(projectInfo.getBackupList);
 
 router.route('/project/:id/editor')
     .get(projectInfo.getProjectById);
+
+router.route('/project/:id/visualization')
+    .get(projectInfo.getProjectTreeById);
 
 router.route('/project/:id/content')
     .get(projectInfo.getProjectContent);